Update PC02_Graffiti_PUDLO.py
First I put a few comments in so i could understand the code.
Set A to 90 to replace having to type 90 in everytime.
Switch the squares to for loops. but when i try this it messes up the position of the squares and i cant figure out why.

Sandwich:
The code was a bit hard to follow without any comments or descriptive turtle names 
I would define all colors pen size and a turtles at the top to keep the code more simple. 
The picture is cool with the gasses and the squares going in a necklace style.